Should now compile.
I found an article that goes into more depth of what the algorithms do, as well as provides a sample program and source code :
http://www.codeproject.com/Articles/644067/Applying-Ant-Colony-Optimization-Algorithms-to-Sol

I'm going to be using the ideas discussed in this article in my program.

Next will be making some big changes to allow use of Travelling salesman problem.  Initially I thought that a grid would work out fine, but in reality, a TSM problem doesn't rely on adjecent cells, it only looks for the paths between cities.

Basically each city has a path to all teh other cities.  I was thinking it was a directed graph with predetermined "roads" so to speak, therefore leading to cities that are only connected by 1 route.

I should be able to reuse a good chunk of the code I currently have, first thing I'm going to do is create a list of cities, along with a cities object.  Most likley I will be renaming the Food class to become a cities class.

Also will be adding a property to the ants class that will allow it to know which cities it has yet to visit.

Thinking that the ant might become more of a powerhouse of the algorithm, since from what I understand, something is keeping track of the where the ant is as well as generic stats about the whole thing.

A couple of things to note :
  initially, each path has some amount of pheromone on them
  ants start on a random city
  ants go from city to city exactly once
  last visited city should connect ot the first city

Ant doesn't necessarily follow the path with teh most pheromone
  diminish chance of a local loop
    don't think I will run into issue with that, since the ant will know what cities are availble for him to go to.
  Pheromone levels don't update until all ants complete a route
    so they do travel concurrently.