All ants should now be able to calculate which city they want to go to next, with the new getNextCity(completeCityList, Alpha, Beta) function.
Not quite sure how effective this is, as it isn't quite like the algorithm described in "Genetic and Ant Algorithms_src"  project.  My particuler implementation is smart enough to know when it has only one city left, in which case it just simply returns the one city without the current algorithm, and when there is none left, in which case it will set the visitedAllCities flag to true, and return the first city.

Might have to rethink this, as the path will need to be updated with values.

Need to find a way to print out the path it choose.