Had to seperate out the initialization of the Ants, they are now required to know the complete list of cities before being created.
This makes it simpler to create properties for the ant object.

Changed teh citiesToVisit function into a property, can now do that due to requiring the pass in of the complete city list.

Ant now knows all the cities he can travel to, and in extension, all the edges as well.  (lstAllCities)