Now have the basics of the drawing of the map.
DrawCity does three things :
  It creates an ellipse with a backgroundcolor of white,
  Then creates another ellipse with a border of blue,
  Finally puts in the city id in the center of the circle.

The reason for the first two steps, is because there is no direct way of setting both the background and the border at the same time.

Also figured out that when drawing these out, the cities need to go last, in order to cover up the lines.  Downside to doing it this way is that the pointing arrows are hidden.

My work around for this right now is by making the first city background color green and the second one yellow.  This won't matter much when just drawing a complete map, but when you start drawing the paths the ants generate, it will always highlight the starting city green and the following city yellow, thus indicating a direction.

Created a new function that will be able to draw the cities with the provided list and canvas. (drawCities)

Also adjusted the city 6 to be a bit higher in the graph, reduces the crowding on the map.

Next step is now making sure that the selection of the next city algorithm works correctly.