Have the basics of the PheromoneUpdate done.  Not quite sure if it will work, as no way of testing it yet.
Should probably update the line to show the pheromone level of each path.

Then need to create another function that will act as the main driver of the whole algorithm, starting with getting the ants moving.

Probably should find some way of compiling the images for each path that the individual ants come up with.

I believe that there should be some way of doing a 100 iterations and comparing the different paths that the ants choose.