AntAlgorithm    This will now be able to have the ants all get their solution.  It will reset the ants current solution everytime it is called. (getAntsSolutions)   There is now a function that will act as a main driver for the algorithm.
The main form will now display two images, one for the before and one for the after the algorithm run.

When ever a path is drawn, the pheromone level will be displayed as well.

Next I will focus on saving the best path for all the ants, by selecting the shortest one of them all, and storing it in a list of some sort.

Will most likely be defaulted to the initial list of cities.

Will be updated with each iteration of the MainDriver()